Updated gender analysis data
